docling-parser-service/requirements.txt
docling
fastapi>=0.115.0,<1.0.0
python-dotenv>=1.0.1,<2.0.0
python-multipart>=0.0.9,<1.0.0
uvicorn>=0.30.0,<1.0.0
